requirements.txt
streamlit
pypdf2
sentence-transformers
faiss-cpu
langchain
langchain-community
langchain-anthropic
python-dotenv
